+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Log failed subscription for submissions (gh-298).
 * Display timezone in the My Subscriptions page (ongoing gh-295)
 * More EU checkboxes since we now have to also deal with GDPR.
+* Google Analytics for e-commerce integration.
 
 **Miscellaneous changes**